// src/app/models/employeeAttendanceLog.ts
export class EmployeeAttendanceLog {
    id?: number;
    employeeId?: number;
    employeeNameFl?:string;
    employeeNameSl?:string;
    employeeNumber?:string;
    terminalSn?:string;
    terminalIp?:string;
    timeEntry?: Date;
    logTypeId?: number;
    remarkId?: number;
    lastModfied?: Date;
    imagePath?: string;
    //timeEntryStr?: string;

}